backend/requirements.txt
# Web Framework
fastapi
uvicorn
python-multipart
email-validator

# Database & Vectors
sqlalchemy
psycopg2-binary
pgvector

# AI Tools
sentence-transformers
groq
google-genai

# Environment Variables
python-dotenv

# Auth (bcrypt directly — passlib breaks on bcrypt 4.1+)
bcrypt>=4.0
python-jose[cryptography]

resend
fpdf2
pdfplumber
pymupdf